move factories to a better place
